@@ -12,7 +12,8 @@
   "dependencies": {
     "browserify": "~9.0.3",
     "literalify": "~0.4.0",
-    "react": "~0.13.1"
+    "react": "^0.14.0",
+    "react-dom": "^0.14.0"
   },
   "devDependencies": {
     "babel": "~5.8.21"
